+++ release/angular-breadcrumb.js
@@ -1,4 +1,4 @@
-/*! angular-breadcrumb - v0.2.3
+/*! angular-breadcrumb - v0.3.0
 * https://github.com/ncuillery/angular-breadcrumb
 * Copyright (c) 2014 Nicolas Cuillery; Licensed MIT */
 
@@ -11,12 +11,19 @@ function isAOlderThanB(scopeA, scopeB) {
     }
 }
 
+function parseStateRef(ref) {
+    var parsed = ref.replace(/\n/g, " ").match(/^([^(]+?)\s*(\((.*)\))?$/);
+    if (!parsed || parsed.length !== 4) { throw new Error("Invalid state ref '" + ref + "'"); }
+    return { state: parsed[1], paramExpr: parsed[3] || null };
+}
+
 function $Breadcrumb() {
 
     var $$options = {
         prefixStateName: null,
         template: 'bootstrap3',
-        templateUrl: null
+        templateUrl: null,
+        includeAbstract : false
     };
 
     this.setOptions = function(options) {
@@ -35,45 +42,52 @@ function $Breadcrumb() {
             }
         });
 
-        // Check if a property in state's data is its own
-        var $$isStateDataProperty = function(state, property) {
-            if(!state.data || !state.data[property]) {
-                return false;
-            }
-
-            var parentState = $$parentState(state);
-            return !(parentState && parentState.data && parentState.data[property] && state.data[property] === parentState.data[property]);
-        };
-
         // Get the parent state
         var $$parentState = function(state) {
             // Check if state has explicit parent OR we try guess parent from its name
             var name = state.parent || (/^(.+)\.[^.]+$/.exec(state.name) || [])[1];
             // If we were able to figure out parent name then get this state
-            return name && $state.get(name);
+            return name;
         };
 
         // Add the state in the chain if not already in and if not abstract
-        var $$addStateInChain = function(chain, state) {
+        var $$addStateInChain = function(chain, stateRef) {
+            var conf,
+                parentParams,
+                ref = parseStateRef(stateRef);
+
             for(var i=0, l=chain.length; i<l; i+=1) {
-              if (chain[i].name === state.name) {
-                return;
-              }
+                if (chain[i].name === ref.state) {
+                    return;
+                }
             }
 
-            if(!state.abstract && !$$isStateDataProperty(state, 'ncyBreadcrumbSkip')) {
-                state.ncyBreadcrumbLink = $state.href(state.name, $stateParams || {});
-                chain.unshift(state);
+            conf = $state.get(ref.state);
+            if((!conf.abstract || $$options.includeAbstract) && !(conf.ncyBreadcrumb && conf.ncyBreadcrumb.skip)) {
+                if(ref.paramExpr) {
+                    parentParams = $lastViewScope.$eval(ref.paramExpr);
+                }
+
+                conf.ncyBreadcrumbLink = $state.href(ref.state, parentParams || $stateParams || {});
+                chain.unshift(conf);
             }
         };
 
         // Get the state for the parent step in the breadcrumb
-        var $$breadcrumbParentState = function(state) {
-            if($$isStateDataProperty(state, 'ncyBreadcrumbParent')) {
-                return $state.get(state.data.ncyBreadcrumbParent);
+        var $$breadcrumbParentState = function(stateRef) {
+            var ref = parseStateRef(stateRef),
+                conf = $state.get(ref.state);
+
+            if(conf.ncyBreadcrumb && conf.ncyBreadcrumb.parent) {
+                // Handle the "parent" property of the breadcrumb, override the parent/child relation of the state
+                var isFunction = typeof conf.ncyBreadcrumb.parent === 'function';
+                var parentStateRef = isFunction ? conf.ncyBreadcrumb.parent($lastViewScope) : conf.ncyBreadcrumb.parent;
+                if(parentStateRef) {
+                    return parentStateRef;
+                }
             }
 
-            return $$parentState(state);
+            return $$parentState(conf);
         };
 
         return {
@@ -94,25 +108,28 @@ function $Breadcrumb() {
                 return $$options.templateUrl;
             },
 
-            getStatesChain: function() {
+            getStatesChain: function(exitOnFirst) { // Deliberately undocumented param, see getLastStep
                 var chain = [];
 
                 // From current state to the root
-                for(var state = $state.$current.self; state && state.name !== ''; state=$$breadcrumbParentState(state)) {
-                  $$addStateInChain(chain, state);
+                for(var stateRef = $state.$current.self.name; stateRef; stateRef=$$breadcrumbParentState(stateRef)) {
+                    $$addStateInChain(chain, stateRef);
+                    if(exitOnFirst && chain.length) {
+                        return chain;
+                    }
                 }
 
                 // Prefix state treatment
                 if($$options.prefixStateName) {
-                    var prefixState = $state.get($$options.prefixStateName);
-                    if(!prefixState) {
-                        throw 'Bad configuration : prefixState "' + $$options.prefixStateName + '" unknown';
-                    }
-
-                    $$addStateInChain(chain, prefixState);
+                    $$addStateInChain(chain, $$options.prefixStateName);
                 }
 
                 return chain;
+            },
+
+            getLastStep: function() {
+                var chain = this.getStatesChain(true);
+                return chain.length ? chain[0] : undefined;
             },
 
             $getLastViewScope: function() {
@@ -122,19 +139,48 @@ function $Breadcrumb() {
     }];
 }
 
+var getExpression = function(interpolationFunction) {
+    if(interpolationFunction.expressions) {
+        return interpolationFunction.expressions;
+    } else {
+        var expressions = [];
+        angular.forEach(interpolationFunction.parts, function(part) {
+            if(angular.isFunction(part)) {
+                expressions.push(part.exp);
+            }
+        });
+        return expressions;
+    }
+};
+
+var registerWatchers = function(labelWatcherArray, interpolationFunction, viewScope, step) {
+    angular.forEach(getExpression(interpolationFunction), function(expression) {
+        var watcher = viewScope.$watch(expression, function() {
+            step.ncyBreadcrumbLabel = interpolationFunction(viewScope);
+        });
+        labelWatcherArray.push(watcher);
+    });
+
+};
+
+var deregisterWatchers = function(labelWatcherArray) {
+    angular.forEach(labelWatcherArray, function(deregisterWatch) {
+        deregisterWatch();
+    });
+    labelWatcherArray = [];
+};
+
 function BreadcrumbDirective($interpolate, $breadcrumb, $rootScope) {
     this.$$templates = {
         bootstrap2: '<ul class="breadcrumb">' +
-            '<li ng-repeat="step in steps | limitTo:(steps.length-1)">' +
-            '<a href="{{step.ncyBreadcrumbLink}}">{{step.ncyBreadcrumbLabel}}</a> ' +
-            '<span class="divider">/</span>' +
-            '</li>' +
-            '<li ng-repeat="step in steps | limitTo:-1" class="active">' +
-            '<span>{{step.ncyBreadcrumbLabel}}</span>' +
+            '<li ng-repeat="step in steps" ng-switch="$last || !!step.abstract" ng-class="{active: $last}">' +
+            '<a ng-switch-when="false" href="{{step.ncyBreadcrumbLink}}">{{step.ncyBreadcrumbLabel}}</a> ' +
+            '<span ng-switch-when="true">{{step.ncyBreadcrumbLabel}}</span>' +
+            '<span class="divider" ng-hide="$last">/</span>' +
             '</li>' +
             '</ul>',
         bootstrap3: '<ol class="breadcrumb">' +
-            '<li ng-repeat="step in steps" ng-class="{active: $last}" ng-switch="$last">' +
+            '<li ng-repeat="step in steps" ng-class="{active: $last}" ng-switch="$last || !!step.abstract">' +
             '<a ng-switch-when="false" href="{{step.ncyBreadcrumbLink}}">{{step.ncyBreadcrumbLabel}}</a> ' +
             '<span ng-switch-when="true">{{step.ncyBreadcrumbLabel}}</span>' +
             '</li>' +
@@ -151,47 +197,16 @@ function BreadcrumbDirective($interpolate, $breadcrumb, $rootScope) {
             post: function postLink(scope) {
                 var labelWatchers = [];
 
-                var getExpression = function(interpolationFunction) {
-                    if(interpolationFunction.expressions) {
-                        return interpolationFunction.expressions;
-                    } else {
-                        var expressions = [];
-                        angular.forEach(interpolationFunction.parts, function(part) {
-                            if(angular.isFunction(part)) {
-                                expressions.push(part.exp);
-                            }
-                        });
-                        return expressions;
-                    }
-                };
-
-                var registerWatchers = function(interpolationFunction, scope, step) {
-                    angular.forEach(getExpression(interpolationFunction), function(expression) {
-                        var watcher = scope.$watch(expression, function() {
-                            step.ncyBreadcrumbLabel = interpolationFunction(scope);
-                        });
-                        labelWatchers.push(watcher);
-                    });
-
-                };
-
-                var deregisterWatchers = function() {
-                    angular.forEach(labelWatchers, function(deregisterWatch) {
-                        deregisterWatch();
-                    });
-                    labelWatchers = [];
-                };
-
                 var renderBreadcrumb = function() {
-                    deregisterWatchers();
+                    deregisterWatchers(labelWatchers);
                     var viewScope = $breadcrumb.$getLastViewScope();
                     scope.steps = $breadcrumb.getStatesChain();
                     angular.forEach(scope.steps, function (step) {
-                        if (step.data && step.data.ncyBreadcrumbLabel) {
-                            var parseLabel = $interpolate(step.data.ncyBreadcrumbLabel);
+                        if (step.ncyBreadcrumb && step.ncyBreadcrumb.label) {
+                            var parseLabel = $interpolate(step.ncyBreadcrumb.label);
                             step.ncyBreadcrumbLabel = parseLabel(viewScope);
                             // Watcher for further viewScope updates
-                            registerWatchers(parseLabel, viewScope, step);
+                            registerWatchers(labelWatchers, parseLabel, viewScope, step);
                         } else {
                             step.ncyBreadcrumbLabel = step.name;
                         }
@@ -210,7 +225,58 @@ function BreadcrumbDirective($interpolate, $breadcrumb, $rootScope) {
 }
 BreadcrumbDirective.$inject = ['$interpolate', '$breadcrumb', '$rootScope'];
 
+function BreadcrumbLastDirective($interpolate, $breadcrumb, $rootScope) {
+
+    return {
+        restrict: 'A',
+        scope: {},
+        template: '{{ncyBreadcrumbLabel}}',
+        compile: function(cElement, cAttrs) {
+
+            // Override the default template if ncyBreadcrumbLast has a value
+            var template = cElement.attr(cAttrs.$attr.ncyBreadcrumbLast);
+            if(template) {
+                cElement.html(template);
+            }
+
+            return {
+                post: function postLink(scope) {
+                    var labelWatchers = [];
+
+                    var renderLabel = function() {
+                        deregisterWatchers(labelWatchers);
+                        var viewScope = $breadcrumb.$getLastViewScope();
+                        var lastStep = $breadcrumb.getLastStep();
+                        if(lastStep) {
+                            scope.ncyBreadcrumbLink = lastStep.ncyBreadcrumbLink;
+                            if (lastStep.ncyBreadcrumb && lastStep.ncyBreadcrumb.label) {
+                                var parseLabel = $interpolate(lastStep.ncyBreadcrumb.label);
+                                scope.ncyBreadcrumbLabel = parseLabel(viewScope);
+                                // Watcher for further viewScope updates
+                                // Tricky last arg: the last step is the entire scope of the directive !
+                                registerWatchers(labelWatchers, parseLabel, viewScope, scope);
+                            } else {
+                                scope.ncyBreadcrumbLabel = lastStep.name;
+                            }
+                        }
+                    };
+
+                    $rootScope.$on('$viewContentLoaded', function () {
+                        renderLabel();
+                    });
+
+                    // View(s) may be already loaded while the directive's linking
+                    renderLabel();
+                }
+            };
+
+        }
+    };
+}
+BreadcrumbLastDirective.$inject = ['$interpolate', '$breadcrumb', '$rootScope'];
+
 angular.module('ncy-angular-breadcrumb', ['ui.router.state'])
     .provider('$breadcrumb', $Breadcrumb)
-    .directive('ncyBreadcrumb', BreadcrumbDirective);
+    .directive('ncyBreadcrumb', BreadcrumbDirective)
+    .directive('ncyBreadcrumbLast', BreadcrumbLastDirective);
 })(window, window.angular);